requirements.txt
boto==2.49.0
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Django==2.1.9
django-storages==1.7.1
django-widget-tweaks==1.4.5
Flask==1.1.1
google-api-core==1.13.0
google-auth==1.6.3
google-cloud-core==1.0.2
google-cloud-storage==1.16.1
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Pillow==5.4.1
protobuf==3.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
PyMySQL==0.9.3
pytz==2019.1
requests==2.22.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
Werkzeug==0.15.4
matplotlib
tensorflow
scikit-image
keras
cloudstorage
IPython